@title Plot of experimental and simulated profile inventories

@description 'plot' returns two plots corresponding to the experimental and simulated inventories for each profile defined in the input data. Automatic function when you execute radi_start. It creates results folder in the current directory if didn't exist.

Usage

plot(data1, data2, dir1, dir2)

Details

@details This 'plot' function performs several operations to create and save a combined plot displaying two types of inventories: one experimental and one simulated. It first creates a "results" folder within the directory specified by dir2. May delete the inner files of results folder if it already exists. delete previous Then, it reads two text files, _num.txt and _exp.txt, from the directory specified by dir1. Next, it creates dataframes containing the necessary data for plotting. It calculates axis limits and labels using data from data1 and data2. Subsequently, it generates separate plots for the experimental and simulated inventories using ggplot2. These plots are combined into a single plot using the plot_layout() function. The combined plot is saved as a PNG file within the results folder, with a name based on the last ID from data2. Finally, the combined plot is displayed in the console. This function utilizes ggplot2 and dplyr extensively for data manipulation and visualization.

@param data1 Dataframe of numeric vectors. Read input data in CSV format. @param data2 dataframe of numeric vectors. Filtered values of data1 for the current 'ID'. @param dir1 Directory. Temporary working directory where the '_num.txt' and '_exp.txt' files were created. @param dir2 Directory. Current working directory.

@returns Results folder in dir2 with n plots saved as a PNG based on the n "ID" defined in data2. Results summarized TXT file in the results folder.